Clean up DB connection
Yeah, we're hard-coding to a prod config, but at least we can now switch.
Also, dependency inject the DB connection. Right now we're reconnecting
to the DB on each request, which is not ideal. However, since I'm not
sure why we're doing that in the first place, I'm not ready to make the
DB connection fixed yet.